Redirecting to survey list

+++ app/src/main/java/com/ipnetinstitute/csc394/mobile/data/model/BaseModel.kt
@@ -2,5 +2,5 @@ package com.ipnetinstitute.csc394.mobile.data.model
 
 import java.util.*
 
-open class BaseModel(var id:Long, var createDate: Date, var modDate: Date, var modBy:Long) {
+open class BaseModel(var id: Int, var createDate: Date, var modDate: Date, var modBy: Int) {
 }
