+++ app/src/main/java/com/ipnetinstitute/csc394/mobile/data/model/Class.kt
@@ -2,6 +2,6 @@ package com.ipnetinstitute.csc394.mobile.data.model
 
 import java.util.*
 
-class Class(var name:String, id: Long, createDate: Date, modDate: Date, modBy: Long) :
+class Class(var name: String, id: Int, createDate: Date, modDate: Date, modBy: Int) :
     BaseModel(id, createDate, modDate, modBy) {
 }
