+++ app/src/main/java/com/ipnetinstitute/csc394/mobile/data/model/Question.kt
@@ -2,6 +2,6 @@ package com.ipnetinstitute.csc394.mobile.data.model
 
 import java.util.*
 
-class Question(var title:String, id: Long, createDate: Date, modDate: Date, modBy: Long) :
+class Question(var title: String, id: Int, createDate: Date, modDate: Date, modBy: Int) :
     BaseModel(id, createDate, modDate, modBy) {
 }
