+++ app/src/main/java/com/ipnetinstitute/csc394/mobile/data/model/Term.kt
@@ -2,6 +2,13 @@ package com.ipnetinstitute.csc394.mobile.data.model
 
 import java.util.*
 
-class Term(var code:String, var name:String, id: Long, createDate: Date, modDate: Date, modBy: Long) :
+class Term(
+    var code: String,
+    var name: String,
+    id: Int,
+    createDate: Date,
+    modDate: Date,
+    modBy: Int
+) :
     BaseModel(id, createDate, modDate, modBy) {
 }
